+++ src/line_launcher/mod.rs
@@ -1,137 +1,24 @@
 use midir::MidiOutputConnection;
 use rand::Rng;
 use std::sync::{
-    mpsc::{self, Receiver, Sender},
+    mpsc::{Receiver, Sender},
     Arc, Mutex,
 };
-use std::thread::{self, JoinHandle};
 use std::time::{Duration, SystemTime};
-use wmidi::{Channel, MidiMessage, Note, Velocity};
 
-use crate::{BeatNumber, Chord, Line, Progression};
+use crate::{BeatNumber, Line, Progression};
 
-#[derive(Clone, Copy, Debug)]
-enum PlayingState {
-    NotPlaying,
-    Playing {
-        line_index: usize,
-        next_note_index: usize,
-        pitch_offset: i8,
-        has_fired_previous_note_off: bool,
-    },
-}
+mod midi_message_sender;
+use midi_message_sender::MidiMessageSender;
 
-enum ProgressionChordIndexState {
-    HaventStarted,
-    AtChordIndex(usize),
-}
+mod note_off_triggerer;
+use note_off_triggerer::{NoteOffInstruction, NoteOffTriggerer};
 
-struct ProgressionState<'progression> {
-    progression: &'progression Progression,
-    chord_index_state: ProgressionChordIndexState,
-}
+mod playing_state;
+use playing_state::PlayingState;
 
-impl<'progression> ProgressionState<'progression> {
-    pub fn new(progression: &'progression Progression) -> Self {
-        Self {
-            progression,
-            chord_index_state: ProgressionChordIndexState::HaventStarted,
-        }
-    }
-
-    pub fn chord_index(&self) -> usize {
-        if let ProgressionChordIndexState::AtChordIndex(chord_index) = self.chord_index_state {
-            chord_index
-        } else {
-            0
-        }
-    }
-
-    pub fn current_chord(&self) -> &Chord {
-        &self.progression.chords[self.chord_index()]
-    }
-
-    pub fn tick_measure(&mut self) {
-        self.chord_index_state = match self.chord_index_state {
-            ProgressionChordIndexState::HaventStarted => {
-                ProgressionChordIndexState::AtChordIndex(0)
-            }
-            ProgressionChordIndexState::AtChordIndex(chord_index) => {
-                ProgressionChordIndexState::AtChordIndex(
-                    (chord_index + 1) % self.progression.chords.len(),
-                )
-            }
-        }
-    }
-}
-
-const CHANNEL: Channel = Channel::Ch1;
-const VELOCITY: u8 = 100;
-
-struct NoteOffInstruction {
-    note: Note,
-    time: SystemTime,
-    note_index: usize,
-}
-
-struct NoteOffTriggerer {
-    receiver: Receiver<NoteOffInstruction>,
-    midi_message_sender: MidiMessageSender,
-    playing_state: Arc<Mutex<PlayingState>>,
-}
-
-impl NoteOffTriggerer {
-    pub fn new(
-        midi_message_sender: MidiMessageSender,
-        playing_state: Arc<Mutex<PlayingState>>,
-    ) -> (Self, Sender<NoteOffInstruction>) {
-        let (sender, receiver) = mpsc::channel();
-
-        let note_off_triggerer = Self {
-            receiver,
-            midi_message_sender,
-            playing_state,
-        };
-
-        (note_off_triggerer, sender)
-    }
-
-    pub fn listen(self) -> JoinHandle<()> {
-        thread::spawn(move || loop {
-            let note_off_instruction = self.receiver.recv().unwrap();
-
-            let now = SystemTime::now();
-            let from_now = note_off_instruction
-                .time
-                .duration_since(now)
-                .unwrap_or_default();
-
-            spin_sleep::sleep(from_now);
-
-            let mut playing_state = self.playing_state.lock().unwrap();
-
-            match *playing_state {
-                PlayingState::Playing {
-                    has_fired_previous_note_off: false,
-                    line_index,
-                    next_note_index,
-                    pitch_offset,
-                } if next_note_index == note_off_instruction.note_index + 1 => {
-                    self.midi_message_sender
-                        .fire_note_off(note_off_instruction.note);
-
-                    *playing_state = PlayingState::Playing {
-                        line_index,
-                        next_note_index,
-                        pitch_offset,
-                        has_fired_previous_note_off: true,
-                    };
-                }
-                _ => (),
-            }
-        })
-    }
-}
+mod progression_state;
+use progression_state::ProgressionState;
 
 pub struct LineLauncher {
     lines: Vec<Line>,
@@ -253,40 +140,5 @@ impl From<Progression> for LineLauncher {
             lines: Line::all(),
             progression,
         }
-    }
-}
-
-#[derive(Clone)]
-struct MidiMessageSender {
-    output: Arc<Mutex<MidiOutputConnection>>,
-}
-
-impl MidiMessageSender {
-    fn new(output: MidiOutputConnection) -> Self {
-        Self {
-            output: Arc::new(Mutex::new(output)),
-        }
-    }
-
-    fn fire_note_on(&self, note: Note) {
-        self.send_midi_message(MidiMessage::NoteOn(
-            CHANNEL,
-            note,
-            Velocity::from_u8_lossy(VELOCITY),
-        ));
-    }
-
-    fn fire_note_off(&self, note: Note) {
-        self.send_midi_message(MidiMessage::NoteOff(
-            CHANNEL,
-            note,
-            Velocity::from_u8_lossy(VELOCITY),
-        ));
-    }
-
-    fn send_midi_message(&self, midi_message: MidiMessage) {
-        let mut bytes_buffer = vec![0; midi_message.bytes_size()];
-        midi_message.copy_to_slice(&mut bytes_buffer).unwrap();
-        self.output.lock().unwrap().send(&bytes_buffer).unwrap();
     }
 }
